buffer2 を削除
