+++ text_buffer/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod action;
 pub mod buffer;
-pub mod buffer2;
 pub mod caret;
 pub mod editor;
 pub mod key;
